+++ tests/feature_tests/src/steps/report.py
@@ -154,7 +154,7 @@ def change_title_of_non_existing_report(context: Context) -> None:
 @then("the import failed")
 def import_failed(context: Context) -> None:
     """Check the JSON."""
-    assert_equal(400, context.response.status_code)
+    assert_equal(422, context.response.status_code)
     assert_equal("application/json", context.response.headers["Content-Type"])